pyproject.toml
[project]
name = "databricks-genie-bot"
version = "0.1.0"
description = "A Microsoft Teams Bot integrating with Databricks Genie via FastAPI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "aiohttp>=3.8.0",
    "aiohttp-compress>=0.2.0",
    "psutil>=5.9.0",
    "botbuilder-core>=4.15.0",
    "botbuilder-ai>=4.15.0",
    "botframework-connector>=4.15.0",
    "botbuilder-schema>=4.15.0",
    "botbuilder-dialogs>=4.15.0",
    "botbuilder-integration-aiohttp>=4.15.0",
    "databricks-sdk>=0.12.0",
    "azure-identity>=1.15.0",
    "tzdata>=2023.3",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "numpy>=1.24.0",
    "fastapi[standard]>=0.112.0",
    "uvicorn>=0.30.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
